Extract expectation of R66 and R69
